Made the "lead" text editable.
Truthfully, this shouldn't be a feature as this site *has* no login functionality. As a portfolio page, no one should be able to publicly edit any of these text fields.

As such, I've limited it to just the lead text as a little fun showcase of JavaScript at work. This editable change will not remain between page refreshes to avoid caching content on the user's system.